requirements.txt
streamlit
supabase
requests
google-genai
python-dotenv
Pillow
pandas
plotly
